+++ lib/src/properties/clear_button_props.dart
@@ -25,6 +25,5 @@ class ClearButtonProps extends IconButtonProps {
     super.style,
     super.isSelected,
     super.selectedIcon,
-    super.onPressed,
   });
 }