+++ lib/src/properties/dropdown_suffix_props.dart
@@ -1,0 +1,14 @@
+import 'package:dropdown_search/dropdown_search.dart';
+import 'package:flutter/material.dart';
+
+class DropdownSuffixProps {
+  final ClearButtonProps clearButtonProps;
+  final DropdownButtonProps dropdownButtonProps;
+  final TextDirection? direction;
+
+  const DropdownSuffixProps({
+    this.clearButtonProps = const ClearButtonProps(),
+    this.dropdownButtonProps = const DropdownButtonProps(),
+    this.direction = TextDirection.ltr,
+  });
+}